move ICommit to "core" to fix external plugin TS build errors
