+++ typings/gitlog.d.ts
@@ -1,26 +1,34 @@
 declare module 'gitlog' {
-  export interface ICommit {
-    hash: string;
-    authorName?: string;
-    authorEmail?: string;
-    subject: string;
-    rawBody?: string;
-    labels?: string[];
-    files: string[];
-  }
-
   interface IGitlogOptions {
-    repo: string;
-    fields: string[];
-    branch: string;
-    number: number;
-    execOptions: {
-      maxBuffer: number;
+    /**
+     *
+     */
+repo: string;
+    /**
+     *
+     */
+fields: string[];
+    /**
+     *
+     */
+branch: string;
+    /**
+     *
+     */
+number: number;
+    /**
+     *
+     */
+execOptions: {
+      /**
+       *
+       */
+maxBuffer: number;
     };
   }
 
-  export default function gitlog(
+  export default function gitlog<T>(
     options: IGitlogOptions,
-    callback: (err: Error, res: ICommit[]) => void
+    callback: (err: Error, res: T[]) => void
   ): void;
 }
